Counter GUI implemented. Code restructured.
Added another state machine for the counter GUI on the bottom right which keeps track of the current progress.

The GUI code was outsourced to gui.as and is not runnable as a single file, if you could there would be errors that some variables couldn't be found, but since the file is only included that is ok.
I usually don't like it like this, but the main level script would blow up eventually.

The counter GUI has an animation. However the images are loaded only when the level is running, so I couldn't load them all ahead of time in Init because this would save some users the hickup on the first loading of the files.