Level Editor implemented. Big code refactoring.
Level Editor should be implemented fully now. (Save, Add, Delete, Apply, MoveUp, MoveDown)

Code has been refactored and states have been moved to their respective files along with their variables.

Editor and game are using the same rabbit statue, they are hiding and showing it when necessary.

Saving the mod JSON is a bit slow, maybe figure out later why that is the case.
Currently there is a check missing if the mod selected is a local mod or from the workshop (or the workshop directory).
Do not display workshop mods in the level editor.