Several bugfixes.
Bugfix, the preview fade image wasn't resized when the window was resized.
Bugfix, camera fov was not set during editor preview.
Bugfix, editor would not restore state correctly because CloseCustomEditor() was not called when there were no unsaved changes and the gui was closed with the X in the title bar.

Bugfix, camera distance was not 0 when the preview was running resulting in incorrect previews.